2023-09-22 day job

+++ src/main/java/com/portfolio/manager/dto/OrderInProgressDTO.java
@@ -1,4 +1,4 @@
 package com.portfolio.manager.dto;
 
-public record OrderInProgressDTO(String buyOrSell, String securityName, String securityCode, Double ratio) {
+public record OrderInProgressDTO(String buyOrSell, String securityName, String securityCode, Integer ratio) {
 }
